Add files via upload
The goal of this Power BI project is to design and develop an interactive dashboard that provides insightful visualizations and data-driven analytics. The dashboard will be built to help decision-makers explore trends, KPIs (Key Performance Indicators), and patterns in the dataset. This project leverages Power BI’s powerful features to transform raw data into meaningful insights.